Move factories for creating MCNLRData to that class
MCNLRObjective should not be responsible for creating MCNLRData
