+++ alpha/alpha-algo/src/main/java/org/neo4j/gds/ml/nodemodels/multiclasslogisticregression/MultiClassNLRObjective.java
@@ -26,19 +26,13 @@ import org.neo4j.gds.embeddings.graphsage.ddl4j.functions.L2NormSquared;
 import org.neo4j.gds.embeddings.graphsage.ddl4j.functions.MatrixConstant;
 import org.neo4j.gds.embeddings.graphsage.ddl4j.functions.MultiClassCrossEntropyLoss;
 import org.neo4j.gds.embeddings.graphsage.ddl4j.functions.Weights;
-import org.neo4j.gds.embeddings.graphsage.ddl4j.tensor.Matrix;
 import org.neo4j.gds.embeddings.graphsage.ddl4j.tensor.Scalar;
 import org.neo4j.gds.embeddings.graphsage.ddl4j.tensor.Tensor;
-import org.neo4j.gds.embeddings.graphsage.subgraph.LocalIdMap;
 import org.neo4j.gds.ml.Objective;
 import org.neo4j.gds.ml.batch.Batch;
-import org.neo4j.gds.ml.features.BiasFeature;
-import org.neo4j.gds.ml.features.FeatureExtraction;
 import org.neo4j.graphalgo.api.Graph;
 
-import java.util.Collection;
 import java.util.List;
-import java.util.TreeSet;
 
 public class MultiClassNLRObjective implements Objective<MultiClassNLRData> {
 
@@ -57,37 +51,6 @@ public class MultiClassNLRObjective implements Objective<MultiClassNLRData> {
         this.targetProperty = targetProperty;
         this.graph = graph;
         this.penalty = penalty;
-    }
-
-    public static MultiClassNLRData makeData(
-        Collection<String> featureProperties,
-        String targetPropertyKey,
-        Graph graph
-    ) {
-        var classIdMap = makeClassIdMap(graph, targetPropertyKey);
-        var weights = initWeights(graph, classIdMap.size(), featureProperties);
-        return MultiClassNLRData.builder()
-            .classIdMap(classIdMap)
-            .weights(weights)
-            .build();
-    }
-
-    private static LocalIdMap makeClassIdMap(Graph graph, String targetPropertyKey) {
-        var classSet = new TreeSet<Long>();
-        var classIdMap = new LocalIdMap();
-        graph.forEachNode(nodeId -> {
-            classSet.add(graph.nodeProperties(targetPropertyKey).longValue(nodeId));
-            return true;
-        });
-        classSet.forEach(classIdMap::toMapped);
-        return classIdMap;
-    }
-
-    private static Weights<Matrix> initWeights(Graph graph, int numberOfClasses, Collection<String> featureProperties) {
-        var featureExtractors = FeatureExtraction.propertyExtractors(graph, featureProperties);
-        featureExtractors.add(new BiasFeature());
-        var featuresPerClass = FeatureExtraction.featureCount(featureExtractors);
-        return new Weights<>(Matrix.fill(0.0, numberOfClasses, featuresPerClass));
     }
 
     @Override
